:art: Apply fixes for compiler warnings

+++ Psong/Ball.cpp
@@ -13,12 +13,12 @@ Ball::Ball(Paddle& paddle1, Paddle& paddle2)
 void Ball::update(const float dt, const int windowWidth, const int windowHeight) {
 	if (sprite.getPosition().x < 0) {
 		paddle2->incrementScore();
-		sprite.setPosition(windowWidth / 2, windowHeight / 2 - sprite.getGlobalBounds().height / 2);
+		sprite.setPosition(windowWidth / 2.0f, windowHeight / 2.0f - sprite.getGlobalBounds().height / 2.0f);
 	}
 
 	if (sprite.getPosition().x + sprite.getGlobalBounds().width > windowWidth) {
 		paddle1->incrementScore();
-		sprite.setPosition(windowWidth / 2, windowHeight / 2 - sprite.getGlobalBounds().height / 2);
+		sprite.setPosition(windowWidth / 2.0f, windowHeight / 2.0f - sprite.getGlobalBounds().height / 2.0f);
 	}
 
 	if (sprite.getPosition().y < 0 ||
@@ -30,10 +30,10 @@ void Ball::update(const float dt, const int windowWidth, const int windowHeight)
 }
 
 void Ball::reset(const int windowWidth, const int windowHeight) {
-	std::srand(std::time(0));
+	std::srand((unsigned int)std::time(0));
 	int rand = std::rand() % 4 + 1;
 	
-	sprite.setPosition(windowWidth / 2, windowHeight / 2 - sprite.getGlobalBounds().height / 2);
+	sprite.setPosition(windowWidth / 2.0f, windowHeight / 2.0f - sprite.getGlobalBounds().height / 2.0f);
 
 	switch (rand) {
 	case 1:
